아이디 롱 타입으로 수정

+++ app/src/main/java/com/example/guru2/auth/User.kt
@@ -1,12 +1,12 @@
 package com.example.guru2.auth
 
 data class User(
-    val id: Int,
+    val id: Long,               // ✅ Long
     val username: String,
     val password: String,
     val nickname: String,
     val gender: String,
     val age: Int,
     val country: String,
-    val solvedCount: Int   // ⭐ 추가
+    val solvedCount: Int        // 푼 문제 수
 )
